Flytt overlapp-funksjon+tilhørende test og slett kalender-mappen
