--- src/frontend/utils/kalender/index.ts
@@ -1,1 +1,0 @@
-export * from './periode';
