--- src/frontend/utils/kalender/periode.ts
@@ -1,22 +1,0 @@
-import { isAfter, isBefore, isSameDay } from 'date-fns';
-
-import type { IsoDatoString } from '../dato';
-import { isoStringTilDate, isoStringTilDateMedFallback, tidenesEnde } from '../dato';
-
-export const periodeOverlapperMedValgtDato = (
-    periodeFom: IsoDatoString,
-    periodeTom: IsoDatoString | undefined,
-    valgtDato: Date
-) => {
-    const periodeFomDate = isoStringTilDate(periodeFom);
-    const periodeTomDate = isoStringTilDateMedFallback({
-        isoString: periodeTom,
-        fallbackDate: tidenesEnde,
-    });
-
-    return (
-        (isAfter(valgtDato, periodeFomDate) && isBefore(valgtDato, periodeTomDate)) ||
-        isSameDay(valgtDato, periodeFomDate) ||
-        isSameDay(valgtDato, periodeTomDate)
-    );
-};
